Core_Backend_Services/JIT_Generator_Service/requirements.txt
langgraph 
langchain-core 
langchain-groq 
langchain-google-genai 
pydantic 
sentence-transformers 
langgraph
langchain-core
langchain-groq
langchain-google-genai
pydantic
sentence-transformers
python-dotenv
sqlalchemy
psycopg2-binary
fastapi
uvicorn
langsmith
pytest
pytest-asyncio
typing-extensions